chronos/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.8.6
aioredis==2.0.1
pytesseract==0.3.10
Pillow==10.1.0
neo4j==5.14.0
pydantic==2.4.2
